refactor(router): Oursourcing routes
