+++ routes/events.go
@@ -1,0 +1,58 @@
+package routes
+
+import (
+	"net/http"
+	"strconv"
+	"telu-event-apps/models"
+
+	"github.com/gin-gonic/gin"
+)
+
+func getEvents(context *gin.Context) {
+	events, err := models.GetAllEvents()
+	if err != nil {
+		context.JSON(http.StatusInternalServerError, gin.H{"message": "Could not fetch events. Try again later."})
+		return
+	}
+
+	context.JSON(http.StatusOK, events)
+}
+
+func getEvent(context *gin.Context) {
+	eventId, err := strconv.ParseInt(context.Param("id"), 10, 64)
+	if err != nil {
+		context.JSON(http.StatusBadRequest, gin.H{"message": "Invalid event ID"})
+		return
+	}
+
+	event, err := models.GetEventByID(eventId)
+	if err != nil {
+		context.JSON(http.StatusInternalServerError, gin.H{"message": "Could not fetch event. Try again later."})
+		return
+	}
+
+	context.JSON(http.StatusOK, event)
+}
+
+func createEvents(context *gin.Context) {
+	// Extracting incoming JSON data into an Event struct
+	var event models.Event
+	err := context.ShouldBindJSON(&event)
+	event.ID = 1
+	event.UserID = 1
+
+	// Error handling
+	if err != nil {
+		context.JSON(http.StatusBadRequest, gin.H{"message": "Could not create events. Try again later."})
+		return
+	}
+
+	// Save data
+	err = event.Save()
+	if err != nil {
+		context.JSON(http.StatusInternalServerError, gin.H{"message": "Could not save event"})
+		return
+	}
+	// Return success response
+	context.JSON(http.StatusCreated, gin.H{"message": "Event created successfully.", "event": event})
+}
