+++ routes/routes.go
@@ -1,0 +1,11 @@
+package routes
+
+import "github.com/gin-gonic/gin"
+
+
+// *gin.Engine is type gin.Default() return
+func RegisterRoutes(server *gin.Engine) {
+	server.GET("/events", getEvents)
+	server.GET("/events/:id", getEvent)
+	server.POST("/events", createEvents)
+}
